# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "3.0",
    "title": "Shelter API",
    "description": "Api for shelter"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/animals": {
      "get": {
        "tags": [
          "Animals"
        ],
        "description": "",
        "parameters": [
          {
            "name": "shelters_id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/animalDataRegister/{animalId}": {
      "get": {
        "tags": [
          "Animals"
        ],
        "description": "",
        "parameters": [
          {
            "name": "animalId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "shelters_id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/animalDataAdoption/{animalId}": {
      "get": {
        "tags": [
          "Animals"
        ],
        "description": "",
        "parameters": [
          {
            "name": "animalId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "shelters_id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/animalRegistration": {
      "post": {
        "tags": [
          "Animals"
        ],
        "description": "",
        "parameters": [
          {
            "name": "shelters_id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "registerAnimal": {
                  "example": "any"
                },
                "registerPeople": {
                  "example": "any"
                },
                "register": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}